feat: wire initial-conditions outputs into the canonical pipeline
